Convert to typescript the slash commands invite files
