--- app/slashcommands-invite/client/client.js
@@ -1,7 +1,0 @@
-import { slashCommands } from '../../utils';
-
-slashCommands.add('invite', undefined, {
-	description: 'Invite_user_to_join_channel',
-	params: '@username',
-	permission: 'add-user-to-joined-room',
-});
